ajustes generales

+++ views/back/listCategoria.php
@@ -42,9 +42,13 @@
 									}?>
 								</td>
 								<td>
+								<?php if ($row['nombre'] != 'generales' && $row['nombre'] != 'Generales'){?>
 									<a href="eliminarCategoria/<?php echo $row['id']; ?>" >
 										<button type="button" class="text-uppercase btn btn-edit pull-right"> <span class="glyphicon glyphicon-remove"></span></button>
 									</a> 
+								<?php }else{ ?>
+									<button type="button" class="text-uppercase btn btn-edit pull-right inactivo"> <span class="glyphicon glyphicon-remove"></span></button>
+								<?php } ?>
 								</td>
 							</tr>
